Add bold/italic rendering, CJK Extension support, font-aware width, and configurable font settings
- Add fake bold (50% bitmap blend) and fake italic (shear transform) support
- Add font-names-bold and font-names-italic config options for real bold/italic fonts
- Bold+italic combined style uses italic font base + fake bold
- Upgrade text buffer and rendering pipeline from u16 to u32 for CJK Extension B-F support
- Extend glyph cache to 0x30000 entries, bypass FcCharSet for >U+FFFF characters
- Add font-aware character width detection (glyph width >1.3x half-width -> full-width)
- Remove width clamp for non-fixed-width fonts
- Rename binary to fbterm-mod, config file to .fbterm-modrc
- Bump version to 1.8, update package name to fbterm-mod

+++ src/fbshell.h
@@ -40,9 +40,9 @@ private:
 	FbShell();
 	~FbShell();
 
-	virtual void drawChars(CharAttr attr, u16 x, u16 y, u16 w, u16 num, u16 *chars, bool *dws);
+	virtual void drawChars(CharAttr attr, u16 x, u16 y, u16 w, u16 num, u32 *chars, bool *dws);
 	virtual bool moveChars(u16 sx, u16 sy, u16 dx, u16 dy, u16 w, u16 h);
-	virtual void drawCursor(CharAttr attr, u16 x, u16 y, u16 c);
+	virtual void drawCursor(CharAttr attr, u16 x, u16 y, u32 c);
 	virtual void modeChanged(ModeType type);
 	virtual void request(RequestType type, u32 val = 0);
 
@@ -67,7 +67,7 @@ private:
 		}
 		bool showed;
 		u16 x, y;
-		u16 code;
+		u32 code;
 		CharAttr attr;
 	} mCursor;
